@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.1 <3.0.0"
+  sdk: '>=3.2.5 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -47,7 +47,6 @@ dependencies:
   location: ^4.4.0
   geolocator: ^9.0.2
   geocoding: ^2.1.1
-  conditional_builder: ^1.0.2
   shared_preferences: ^2.2.0
   timer_count_down: ^2.2.2
   flutter_timer_countdown: ^1.0.7
