@@ -3,7 +3,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 class CacheHelper
 {
-  static SharedPreferences sharedPreferences;
+  static SharedPreferences? sharedPreferences;
 
   static init() async
   {
@@ -12,36 +12,36 @@ class CacheHelper
   }
 
   static Future<bool> putBoolean({
-    @required String key,
-    @required bool value,
+    required String key,
+    required bool value,
   }) async
   {
-    return await sharedPreferences.setBool(key, value);
+    return await sharedPreferences!.setBool(key, value);
   }
 
   static dynamic getData({
-    @required String key,
+    required String key,
   }) {
     print('getData called');
-    return sharedPreferences.get(key);
+    return sharedPreferences!.get(key);
   }
 
   static Future<bool> saveData({
-    @required String key,
+    required String key,
     @required dynamic value,
   }) async {
     print('saveData called');
-    if (value is String) return await sharedPreferences.setString(key, value);
-    if (value is int) return await sharedPreferences.setInt(key, value);
-    if (value is bool) return await sharedPreferences.setBool(key, value);
+    if (value is String) return await sharedPreferences!.setString(key, value);
+    if (value is int) return await sharedPreferences!.setInt(key, value);
+    if (value is bool) return await sharedPreferences!.setBool(key, value);
 
-    return await sharedPreferences.setDouble(key, value);
+    return await sharedPreferences!.setDouble(key, value);
   }
 
   static Future<bool> removeData({
-    @required String key,
+    required String key,
   }) async
   {
-    return await sharedPreferences.remove(key);
+    return await sharedPreferences!.remove(key);
   }
 }
